lambda-Functions/requirements.txt
bcrypt==3.1.7
cffi==1.14.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
six==1.15.0
